+++ __tests__/clients/inworld.client.spec.ts
@@ -17,7 +17,6 @@ describe('should finish with success', () => {
   const onError = jest.fn();
   const onMessage = jest.fn();
   const onDisconnect = jest.fn();
-  const generateSessionTokenFn = jest.fn();
   const sessionGetterSetter = {
     get: jest.fn(),
     set: jest.fn(),
@@ -34,7 +33,6 @@ describe('should finish with success', () => {
       .setOnDisconnect(onDisconnect)
       .setOnMessage(onMessage)
       .setOnError(onError)
-      .setGenerateSessionToken(generateSessionTokenFn)
       .setOnSession(sessionGetterSetter);
   });