+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Remove deprecated method support for token, emotions, character and client
+
 ## 2023-05-17 v1.4.0
 * Replace CancelResponses by Mutation.CancelResponses on the protocol level
 * Allow to use extended capabilities and scene props